+++ packages/terra-collapsible-menu-view/src/CollapsibleMenuViewItemGroup.jsx
@@ -4,6 +4,7 @@ import classNames from 'classnames/bind';
 import ButtonGroup from 'terra-button-group';
 import Menu from './_CollapsibleMenu';
 import styles from './CollapsibleMenuView.module.scss';
+import CollapsibleMenuView from '../lib/CollapsibleMenuView';
 
 const cx = classNames.bind(styles);
 
@@ -20,6 +21,10 @@ const propTypes = {
    * A list of keys of the CollapsibleMenuView.Items that should be selected.
    */
   selectedKeys: PropTypes.arrayOf(PropTypes.string),
+  /**
+   * Indicates if the item group should be multi select.
+   */
+  isMultiSelect: PropTypes.bool,
 };
 
 const defaultProps = {
@@ -38,6 +43,8 @@ class CollapsibleMenuViewItemGroup extends React.Component {
   constructor(props) {
     super(props);
     this.handleOnChange = this.handleOnChange.bind(this);
+    this.handleDisplayTypeChange = this.handleDisplayTypeChange.bind(this);
+    this.state = { updatedKeys: this.props.selectedKeys };
   }
 
   getChildContext() {
@@ -57,17 +64,32 @@ class CollapsibleMenuViewItemGroup extends React.Component {
     }
   }
 
+  handleDisplayTypeChange(event, selectedKey) {
+    if (this.props.isMultiSelect) {
+      event.preventDefault();
+      this.setState(prevState => ({ updatedKeys: CollapsibleMenuView.Utils.handleMultiSelectedKeys(prevState.updatedKeys, selectedKey) }), () => {
+        this.props.onChange(event, this.state.updatedKeys);
+      });
+    } else if (CollapsibleMenuView.Utils.shouldHandleSingleSelection(this.state.updatedKeys, selectedKey)) {
+      event.preventDefault();
+      this.setState({ updatedKeys: [selectedKey] }, () => {
+        this.props.onChange(event, this.state.updatedKeys);
+      });
+    }
+  }
+
   render() {
     const {
       children,
       onChange,
       selectedKeys,
+      isMultiSelect,
       ...customProps
     } = this.props;
 
     const { isCollapsibleMenuItem } = this.context;
 
-    if (isCollapsibleMenuItem && selectedKeys.length) {
+    if (isCollapsibleMenuItem && selectedKeys.length && !isMultiSelect) {
       return (
         <li role="none">
           <Menu.ItemGroup {...customProps} onChange={this.handleOnChange}>
@@ -75,7 +97,25 @@ class CollapsibleMenuViewItemGroup extends React.Component {
           </Menu.ItemGroup>
         </li>
       );
-    } if (isCollapsibleMenuItem) {
+    }
+
+    if (isCollapsibleMenuItem && selectedKeys.length && isMultiSelect) {
+      return (
+        children.map((child) => (
+          <Menu.Item
+            text={child.props.text}
+            isSelected={child.props.isSelected}
+            isDisabled={child.props.isDisabled}
+            isIconOnly={child.props.isIconOnly}
+            index={child.props.index}
+            totalItems={child.props.totalItems}
+            isToggleable={isMultiSelect ? true : undefined}
+          />
+        ))
+      );
+    }
+
+    if (isCollapsibleMenuItem) {
       return (
         <div {...customProps}>
           { children }
@@ -89,7 +129,7 @@ class CollapsibleMenuViewItemGroup extends React.Component {
     ]);
 
     return (
-      <ButtonGroup {...customProps} onChange={onChange} className={buttonGroupClassNames} selectedKeys={selectedKeys}>
+      <ButtonGroup {...customProps} onChange={this.handleDisplayTypeChange} isMultiSelect={isMultiSelect} className={buttonGroupClassNames} selectedKeys={selectedKeys}>
         {children}
       </ButtonGroup>
     );