+++ packages/terra-framework-docs/src/terra-dev-site/test/collapsible-menu-view/CollapsibleMenuViewStartAlignment.test.jsx
@@ -20,7 +20,7 @@ class CollapsibleMenuViewDemo extends React.Component {
       toggle1Selection: false,
       toggle2Selection: false,
       toggle3Selection: false,
-      displayType: 'tableView',
+      displayType: ['tableView', 'expandedView'],
       buttonClickCount: 0,
     };
   }
@@ -105,33 +105,37 @@ class CollapsibleMenuViewDemo extends React.Component {
           ]}
         />
         <CollapsibleMenuView.Divider key="Divider2" />
-        <CollapsibleMenuView.ItemGroup key="ViewTypeSelection" className="ViewTypeGroup" selectedKeys={[this.state.displayType]} onChange={this.handleDisplayTypeChange}>
+        <CollapsibleMenuView.ItemGroup key="ViewTypeSelection" isMultiSelect selectedKeys={this.state.displayType} onChange={this.handleDisplayTypeChange}>
           <CollapsibleMenuView.Item
             icon={<IconTable />}
             text="Table View"
-            className="TableView"
             key="tableView"
+            className="TableView"
             isIconOnly
             shouldCloseOnClick={false}
-            isSelected={this.state.displayType === 'tableView'}
+            isSelected={(this.state.displayType).includes('tableView')}
+            isMultiSelect
           />
           <CollapsibleMenuView.Item
             icon={<IconFlowsheet />}
             text="Expanded View"
-            className="ExpandedView"
             key="expandedView"
+            className="ExpandedView"
             isIconOnly
             shouldCloseOnClick={false}
-            isSelected={this.state.displayType === 'expandedView'}
+            isSelected={(this.state.displayType).includes('expandedView')}
+            isMultiSelect
           />
           <CollapsibleMenuView.Item
             icon={<IconVisualization />}
             text="Trending View"
-            className="TendingView"
             key="trendingView"
+            className="TrendingView"
             isIconOnly
             shouldCloseOnClick={false}
-            isSelected={this.state.displayType === 'trendingView'}
+            isSelected={(this.state.displayType).includes('trendingView')}
+            isDisabled
+            isMultiSelect
           />
         </CollapsibleMenuView.ItemGroup>
         <CollapsibleMenuView.Divider key="Divider3" />